[add] 폴더 명 변경 [#42]
