+++ src/main/resources/aws.yml
@@ -1,0 +1,12 @@
+cloud:
+  aws:
+    credentials:
+      accessKey: 
+      secretKey:
+    s3:
+      bucket: dangdang-bucket
+    region:
+      static: ap-northeast-2
+    stack:
+      auto: false
+
